@segmentio/infra 
@sperand-io 

Follow up of https://github.com/segmentio/go-snakecase/pull/1

Before:
```
$ go test -bench .
PASS
BenchmarkCamelcase-4	  500000	      2490 ns/op
ok  	github.com/segmentio/go-camelcase	1.280s
```

After:
```
$ go test -bench .
PASS
BenchmarkCamelcase-4	10000000	       192 ns/op
ok  	github.com/segmentio/go-camelcase	2.126s
```

==> 12x 💃